Now health bar has red effect for damage + deleted commented line
